# v2/api/openapi/public.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "agentserver v2 user resource API",
    "version": "0.1.0",
    "description": "Platform and Browser user resources. Core requires the matching gateway workload identity and validates every original opaque user bearer through Hydra introspection, including its exact client, audience, action scopes, and workspace resource grant."
  },
  "servers": [{"url": "https://agentserver-core"}],
  "security": [{"browserGatewayMTLS": [], "browserOAuth": ["runs:read"]}],
  "paths": {
    "/v2/workspaces": {
      "get": {
        "operationId": "listWorkspaces",
        "description": "Lists the active user's persisted workspace memberships using global Platform read authority.",
        "security": [{"platformGatewayMTLS": [], "platformOAuth": ["workspaces:read"]}],
        "responses": {
          "200": {"description": "Granted workspaces", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ListWorkspacesResponse"}}}},
          "401": {"$ref": "#/components/responses/PublicError"},
          "503": {"$ref": "#/components/responses/PublicError"}
        }
      },
      "post": {
        "operationId": "createWorkspace",
        "description": "Creates an active workspace and its first owner membership in one transaction.",
        "security": [{"platformGatewayMTLS": [], "platformOAuth": ["workspaces:create"]}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateWorkspaceRequest"}}}},
        "responses": {
          "200": {"description": "Exact workspace already exists", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateWorkspaceResponse"}}}},
          "201": {"description": "Workspace created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateWorkspaceResponse"}}}},
          "400": {"$ref": "#/components/responses/PublicError"},
          "401": {"$ref": "#/components/responses/PublicError"},
          "403": {"$ref": "#/components/responses/PublicError"},
          "409": {"$ref": "#/components/responses/PublicError"}
        }
      }
    },
    "/v2/workspaces/{workspaceId}": {
      "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}],
      "get": {
        "operationId": "getWorkspace",
        "security": [{"platformGatewayMTLS": [], "platformOAuth": ["workspaces:read"]}],
        "responses": {
          "200": {"description": "Workspace state", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkspaceState"}}}},
          "401": {"$ref": "#/components/responses/PublicError"},
          "404": {"$ref": "#/components/responses/PublicError"}
        }
      },
      "patch": {
        "operationId": "updateWorkspace",
        "security": [{"platformGatewayMTLS": [], "platformOAuth": ["workspaces:update"]}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateWorkspaceRequest"}}}},
        "responses": {
          "200": {"description": "Updated workspace", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateWorkspaceResponse"}}}},
          "400": {"$ref": "#/components/responses/PublicError"},
          "401": {"$ref": "#/components/responses/PublicError"},
          "403": {"$ref": "#/components/responses/PublicError"},
          "404": {"$ref": "#/components/responses/PublicError"},
          "409": {"$ref": "#/components/responses/PublicError"}
        }
      }
    },
    "/v2/workspaces/{workspaceId}/actions/archive": {
      "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}],
      "post": {
        "operationId": "archiveWorkspace",
        "security": [{"platformGatewayMTLS": [], "platformOAuth": ["workspaces:archive"]}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ArchiveWorkspaceRequest"}}}},
        "responses": {
          "200": {"description": "Archived workspace", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ArchiveWorkspaceResponse"}}}},
          "400": {"$ref": "#/components/responses/PublicError"},
          "401": {"$ref": "#/components/responses/PublicError"},
          "403": {"$ref": "#/components/responses/PublicError"},
          "404": {"$ref": "#/components/responses/PublicError"},
          "409": {"$ref": "#/components/responses/PublicError"}
        }
      }
    },
    "/v2/workspaces/{workspaceId}/members": {
      "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}],
      "get": {
        "operationId": "listWorkspaceMembers",
        "security": [{"platformGatewayMTLS": [], "platformOAuth": ["members:read"]}],
        "responses": {
          "200": {"description": "Active workspace members", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ListWorkspaceMembersResponse"}}}},
          "401": {"$ref": "#/components/responses/PublicError"},
          "403": {"$ref": "#/components/responses/PublicError"},
          "404": {"$ref": "#/components/responses/PublicError"}
        }
      },
      "post": {
        "operationId": "addWorkspaceMember",
        "security": [{"platformGatewayMTLS": [], "platformOAuth": ["members:add"]}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AddWorkspaceMemberRequest"}}}},
        "responses": {
          "200": {"description": "Exact member already exists", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AddWorkspaceMemberResponse"}}}},
          "201": {"description": "Member added", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AddWorkspaceMemberResponse"}}}},
          "400": {"$ref": "#/components/responses/PublicError"},
          "401": {"$ref": "#/components/responses/PublicError"},
          "403": {"$ref": "#/components/responses/PublicError"},
          "404": {"$ref": "#/components/responses/PublicError"},
          "409": {"$ref": "#/components/responses/PublicError"}
        }
      }
    },
    "/v2/workspaces/{workspaceId}/members/{memberId}": {
      "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}, {"$ref": "#/components/parameters/MemberId"}],
      "patch": {
        "operationId": "updateWorkspaceMember",
        "security": [{"platformGatewayMTLS": [], "platformOAuth": ["members:update"]}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateWorkspaceMemberRequest"}}}},
        "responses": {
          "200": {"description": "Updated member", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateWorkspaceMemberResponse"}}}},
          "400": {"$ref": "#/components/responses/PublicError"},
          "401": {"$ref": "#/components/responses/PublicError"},
          "403": {"$ref": "#/components/responses/PublicError"},
          "404": {"$ref": "#/components/responses/PublicError"},
          "409": {"$ref": "#/components/responses/PublicError"}
        }
      },
      "delete": {
        "operationId": "removeWorkspaceMember",
        "security": [{"platformGatewayMTLS": [], "platformOAuth": ["members:remove"]}],
        "responses": {
          "200": {"description": "Member removal result", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RemoveWorkspaceMemberResponse"}}}},
          "400": {"$ref": "#/components/responses/PublicError"},
          "401": {"$ref": "#/components/responses/PublicError"},
          "403": {"$ref": "#/components/responses/PublicError"},
          "409": {"$ref": "#/components/responses/PublicError"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "browserGatewayMTLS": {"type": "mutualTLS", "description": "Exact configured browser-gateway SPIFFE workload identity."},
      "platformGatewayMTLS": {"type": "mutualTLS", "description": "Exact configured platform-gateway SPIFFE workload identity."},
      "browserOAuth": {
        "type": "oauth2",
        "description": "Original agentserver-browser token. Core requires active introspection, exact client_id, single aud=agentserver-browser-api, and one workspace resource grant matching the request path.",
        "flows": {"authorizationCode": {"authorizationUrl": "https://auth-sg.byted.bps.dev/oauth2/auth", "tokenUrl": "https://auth-sg.byted.bps.dev/oauth2/token", "scopes": {
          "sessions:read": "List and read the current user's sessions",
          "sessions:create": "Create a session for the current user",
          "sessions:update": "Rename the current user's session",
          "sessions:archive": "Archive the current user's session",
          "runs:read": "Read the current user's run state and committed event projections",
          "runs:create": "Create a run in the current user's session",
          "runs:cancel": "Cancel the current user's run",
          "approvals:decide": "Approve or deny a live approval for the current user's run"
        }}}
      },
      "platformOAuth": {
        "type": "oauth2",
        "description": "Original agentserver-platform token. Core requires active introspection, exact client_id, single aud=agentserver-platform-api, and a matching global or workspace management grant.",
        "flows": {"authorizationCode": {"authorizationUrl": "https://auth-sg.byted.bps.dev/oauth2/auth", "tokenUrl": "https://auth-sg.byted.bps.dev/oauth2/token", "scopes": {
          "workspaces:read": "List and read granted workspaces",
          "workspaces:create": "Create a workspace",
          "workspaces:update": "Update a granted workspace",
          "workspaces:archive": "Archive a granted workspace",
          "members:read": "Read workspace members",
          "members:add": "Add a workspace member",
          "members:update": "Change a workspace member role",
          "members:remove": "Remove a workspace member",
          "executors:read": "Read workspace executor state",
          "executors:create": "Create a workspace executor identity",
          "executors:enroll": "Issue a workspace executor enrollment token",
          "executors:update": "Update workspace executor metadata",
          "executors:archive": "Archive a workspace executor",
          "llm-gateways:read": "Read workspace LLM Gateway configuration and the current user's grant status",
          "llm-gateways:create": "Create a workspace LLM Gateway",
          "llm-gateways:disable": "Disable a workspace LLM Gateway",
          "llm-gateway-grants:authorize": "Authorize the current user's third-party Gateway grant",
          "llm-gateway-grants:revoke": "Revoke the current user's third-party Gateway grant"
        }}}
      }
...
    "parameters": {
...
      "MemberId": {"name": "memberId", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/UUID"}},
      "WorkspaceId": {"name": "workspaceId", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/UUID"}},
...
    },
    "responses": {
      "PublicError": {"description": "Bounded public error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PublicErrorResponse"}}}}
    },
    "schemas": {
      "UUID": {"type": "string", "format": "uuid", "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$", "not": {"const": "00000000-0000-0000-0000-000000000000"}},
      "WorkspaceState": {
        "type": "object", "additionalProperties": false,
        "required": ["workspaceId", "name", "status", "currentUserRole", "version", "createdAt", "updatedAt"],
        "properties": {
          "workspaceId": {"$ref": "#/components/schemas/UUID"},
          "name": {"type": "string", "minLength": 1, "maxLength": 256},
          "status": {"enum": ["active", "suspended", "archived"]},
          "currentUserRole": {"enum": ["owner", "developer", "viewer"]},
          "version": {"type": "integer", "minimum": 1, "maximum": 9007199254740991},
          "createdAt": {"type": "string", "format": "date-time"},
          "updatedAt": {"type": "string", "format": "date-time"}
        }
      },
      "ListWorkspacesResponse": {
        "type": "object", "additionalProperties": false, "required": ["workspaces"],
        "properties": {"workspaces": {"type": "array", "maxItems": 256, "items": {"$ref": "#/components/schemas/WorkspaceState"}}}
      },
      "CreateWorkspaceRequest": {
        "type": "object", "additionalProperties": false, "required": ["workspaceId", "name"],
        "properties": {"workspaceId": {"$ref": "#/components/schemas/UUID"}, "name": {"type": "string", "minLength": 1, "maxLength": 256}}
      },
      "CreateWorkspaceResponse": {
        "type": "object", "additionalProperties": false, "required": ["workspace", "created"],
        "properties": {"workspace": {"$ref": "#/components/schemas/WorkspaceState"}, "created": {"type": "boolean"}}
      },
      "UpdateWorkspaceRequest": {
        "type": "object", "additionalProperties": false, "required": ["name", "expectedVersion"],
        "properties": {"name": {"type": "string", "minLength": 1, "maxLength": 256}, "expectedVersion": {"type": "integer", "minimum": 1, "maximum": 9007199254740991}}
      },
      "UpdateWorkspaceResponse": {
        "type": "object", "additionalProperties": false, "required": ["workspace", "changed"],
        "properties": {"workspace": {"$ref": "#/components/schemas/WorkspaceState"}, "changed": {"type": "boolean"}}
      },
      "ArchiveWorkspaceRequest": {
        "type": "object", "additionalProperties": false, "required": ["expectedVersion"],
        "properties": {"expectedVersion": {"type": "integer", "minimum": 1, "maximum": 9007199254740991}}
      },
      "ArchiveWorkspaceResponse": {
        "type": "object", "additionalProperties": false, "required": ["workspace", "changed"],
        "properties": {"workspace": {"$ref": "#/components/schemas/WorkspaceState"}, "changed": {"type": "boolean"}}
      },
      "WorkspaceMemberState": {
        "type": "object", "additionalProperties": false, "required": ["userId", "role", "version", "createdAt", "updatedAt"],
        "properties": {
          "userId": {"$ref": "#/components/schemas/UUID"}, "role": {"enum": ["owner", "developer", "viewer"]},
          "version": {"type": "integer", "minimum": 1, "maximum": 9007199254740991},
          "createdAt": {"type": "string", "format": "date-time"}, "updatedAt": {"type": "string", "format": "date-time"}
        }
      },
      "ListWorkspaceMembersResponse": {
        "type": "object", "additionalProperties": false, "required": ["members"],
        "properties": {"members": {"type": "array", "maxItems": 256, "items": {"$ref": "#/components/schemas/WorkspaceMemberState"}}}
      },
      "AddWorkspaceMemberRequest": {
        "type": "object", "additionalProperties": false, "required": ["userId", "role"],
        "properties": {"userId": {"$ref": "#/components/schemas/UUID"}, "role": {"enum": ["owner", "developer", "viewer"]}}
      },
      "AddWorkspaceMemberResponse": {
        "type": "object", "additionalProperties": false, "required": ["member", "created"],
        "properties": {"member": {"$ref": "#/components/schemas/WorkspaceMemberState"}, "created": {"type": "boolean"}}
      },
      "UpdateWorkspaceMemberRequest": {
        "type": "object", "additionalProperties": false, "required": ["role", "expectedVersion"],
        "properties": {"role": {"enum": ["owner", "developer", "viewer"]}, "expectedVersion": {"type": "integer", "minimum": 1, "maximum": 9007199254740991}}
      },
      "UpdateWorkspaceMemberResponse": {
        "type": "object", "additionalProperties": false, "required": ["member", "changed"],
        "properties": {"member": {"$ref": "#/components/schemas/WorkspaceMemberState"}, "changed": {"type": "boolean"}}
      },
      "RemoveWorkspaceMemberResponse": {
        "type": "object", "additionalProperties": false, "required": ["userId", "removed"],
        "properties": {"userId": {"$ref": "#/components/schemas/UUID"}, "removed": {"type": "boolean"}}
      },
...
      "PublicErrorResponse": {
        "type": "object", "additionalProperties": false, "required": ["code", "message"],
        "properties": {
          "code": {"type": "string", "minLength": 1, "maxLength": 128, "pattern": "^[a-z0-9_]+$"},
          "message": {"type": "string", "minLength": 1, "maxLength": 1024},
          "currentRunId": {"$ref": "#/components/schemas/UUID"}
        }
      }
    }
  },
...
}